# packages/contracts/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "description": "API for camp registration and camp operations.",
    "title": "Camp Registration API",
    "version": "0.0.0"
  },
...
  "paths": {
    "/v1/catalog": {
      "get": {
        "tags": [
          "catalog"
        ],
        "description": "Catalog organization, season, and program options.",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "organization",
                    "seasons",
                    "programs"
                  ],
                  "properties": {
                    "organization": {
                      "additionalProperties": false,
                      "type": "object",
                      "required": [
                        "id",
                        "slug",
                        "name",
                        "timezone"
                      ],
                      "properties": {
                        "id": {
                          "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[1-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
                          "type": "string"
                        },
                        "slug": {
                          "minLength": 1,
                          "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
                          "type": "string"
                        },
                        "name": {
                          "minLength": 1,
                          "type": "string"
                        },
                        "timezone": {
                          "minLength": 1,
                          "type": "string"
                        }
                      }
                    },
                    "seasons": {
                      "type": "array",
                      "items": {
                        "additionalProperties": false,
                        "type": "object",
                        "required": [
                          "id",
                          "organization_id",
                          "name",
                          "year"
                        ],
                        "properties": {
                          "id": {
                            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[1-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
                            "type": "string"
                          },
                          "organization_id": {
                            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[1-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
                            "type": "string"
                          },
                          "name": {
                            "minLength": 1,
                            "type": "string"
                          },
                          "year": {
                            "minimum": 2000,
                            "maximum": 2200,
                            "type": "integer"
                          }
                        }
                      }
                    },
                    "programs": {
                      "type": "array",
                      "items": {
                        "additionalProperties": false,
                        "type": "object",
                        "required": [
                          "id",
                          "organization_id",
                          "code",
                          "name",
                          "delivery_mode",
                          "description"
                        ],
                        "properties": {
                          "id": {
                            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[1-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
                            "type": "string"
                          },
                          "organization_id": {
                            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[1-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
                            "type": "string"
                          },
                          "code": {
                            "minLength": 1,
                            "pattern": "^[A-Z0-9-]+$",
                            "type": "string"
                          },
                          "name": {
                            "minLength": 1,
                            "type": "string"
                          },
                          "delivery_mode": {
                            "anyOf": [
                              {
                                "type": "string",
                                "enum": [
                                  "DAY"
                                ]
                              },
                              {
                                "type": "string",
                                "enum": [
                                  "OVERNIGHT"
                                ]
                              }
                            ]
                          },
                          "description": {
                            "minLength": 1,
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "message": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "field_errors": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "message": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "field_errors": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "message": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "field_errors": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "message": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "field_errors": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "message": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "field_errors": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/seasons": {
      "post": {
        "tags": [
          "seasons"
        ],
        "description": "Create a season for the active organization.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "type": "object",
                "required": [
                  "name",
                  "year"
                ],
                "properties": {
                  "name": {
                    "minLength": 1,
                    "maxLength": 160,
                    "type": "string"
                  },
                  "year": {
                    "minimum": 2000,
                    "maximum": 2200,
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "id",
                    "organization_id",
                    "name",
                    "year"
                  ],
                  "properties": {
                    "id": {
                      "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[1-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
                      "type": "string"
                    },
                    "organization_id": {
                      "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[1-8][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
                      "type": "string"
                    },
                    "name": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "year": {
                      "minimum": 2000,
                      "maximum": 2200,
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "message": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "field_errors": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "message": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "field_errors": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "message": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "field_errors": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "message": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "field_errors": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "message": {
                      "minLength": 1,
                      "type": "string"
                    },
                    "field_errors": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "description": "Local development",
      "url": "http://localhost:3001"
    }
  ]
}